+++ app/controllers/frontend/bots_controller.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Frontend
+  class BotsController < Frontend::BaseController
+    include Deps[
+      handle_service: 'services.bot_context.handle'
+    ]
+
+    def create
+      render json: {
+        result: handle_service.call(source: :web, message: params[:value], data: { user: current_user })
+      }, status: :ok
+    end
+  end
+end